SecurityOptions init moved to SecurityOptionsCreator
