+++ src/Ocelot/Configuration/Creator/SecurityOptionsCreator.cs
@@ -1,4 +1,7 @@
 ﻿using Ocelot.Configuration.File;
+using System.Linq;
+using System.Collections.Generic;
+using IPAddressRange = NetTools;
 
 namespace Ocelot.Configuration.Creator
 {
@@ -6,7 +9,33 @@ namespace Ocelot.Configuration.Creator
     {
         public SecurityOptions Create(FileSecurityOptions securityOptions)
         {
-            return new SecurityOptions(securityOptions.IPAllowedList, securityOptions.IPBlockedList, securityOptions.ExcludeAllowedFromBlocked);
+            var IPAllowedList = new List<string>();
+            var IPBlockedList = new List<string>();
+
+            foreach (var allowed in securityOptions.IPAllowedList)
+            {
+                if (IPAddressRange.IPAddressRange.TryParse(allowed, out var allowedIpAddressRange))
+                {
+                    var allowedIps = allowedIpAddressRange.AsEnumerable().Select(x => x.ToString());
+                    IPAllowedList.AddRange(allowedIps);
+                }
+            }
+
+            foreach (var blocked in securityOptions.IPBlockedList)
+            {
+                if (IPAddressRange.IPAddressRange.TryParse(blocked, out var blockedIpAddressRange))
+                {
+                    var blockedIps = blockedIpAddressRange.AsEnumerable().Select(x => x.ToString());
+                    IPBlockedList.AddRange(blockedIps);
+                }
+            }
+
+            if (securityOptions.ExcludeAllowedFromBlocked)
+            {
+                IPBlockedList = IPBlockedList.Except(IPAllowedList).ToList();
+            }
+
+            return new SecurityOptions(IPAllowedList, IPBlockedList);
         }
     }
 }
